tests: Add a test for sync with potential conflicts
This is an scenario where GUW sync command raises some git conflicts that
are avoidable:

	* 0632076 (HEAD -> A) foo
	| * b1ac2a0 (B) bar
	| * fd1dd3f foo
	|/
	* 5ccd4f9 (main, final) init

In this scenario the branch A was rebased, hypothetically because a
reviewed asked for a change. Then, when we try to "sync" with GUW, it
launches the command:

	git rebase origin/A B --onto=final

The range "origin/A..B" is wrong, the correct one should be "B~..B" because
it includes **only** the commits within the "B" branch. This command
doesn't raises conflicts and gives the expected result:

	git rebase B~ B --onto=final